Just committing so I can work on this at home.
Please ignore this commit (please)

Seriously, working on a launcher for a Resnet container. Launching and
cleaning up the container is working nicely. I'm trying to get a stream
object of the container log so we can check on a container's status
while it's running, and that's givine me trouble.

Once that's figured out, I'm going to try to get sysstat working and
collecting data during the runs.