Big overhaul
We can now assign inputs and outputs of the whole function into memory,
or opt out of the whole thing altogether.

Moving a lot of code to a more multiple dispatch oriented design with
resulted in a lot of nice behaviors.

Querying the state of a solution is not much easier. Now onto profiling
and understanding what the heck the ILP is doing!!

Also, I broke synchronous because it's not immediately clear to my tired
brain how to handle some corner cases at the moment.